requirements.txt
beautifulsoup4
pandas
pycryptodome
Requests
selenium
openpyxl
lxml